refactor: components in new maps #122
