+++ src/lib/organisms/DetailsMain.svelte
