+++ src/lib/organisms/FooterInfo.svelte
