+++ src/lib/organisms/SponsorCarousel.svelte
